deleted stupid C++ typecasts.  LAME is written in C. integer arguments to fwrite() do not need typecasts.

+++ frontend/main.c
@@ -20,7 +20,7 @@
  * Boston, MA 02111-1307, USA.
  */
 
-/* $Id: main.c,v 1.36 2000/12/11 16:03:06 aleidinger Exp $ */
+/* $Id: main.c,v 1.37 2000/12/19 07:00:09 markt Exp $ */
 
 #ifdef HAVE_CONFIG_H
 # include <config.h>
@@ -289,8 +289,7 @@ int main(int argc, char **argv)
 	  return 1;
 	}
 
-	/* imp3 is not negative, but fwrite needs an unsigned here */
-	if (fwrite(mp3buffer,1,(unsigned int)imp3,outf) != (size_t)imp3) {
+	if (fwrite(mp3buffer,1,imp3,outf) != imp3) {
 	  fprintf(stderr,"Error writing mp3 output \n");
 	  return 1;
 	}
@@ -317,8 +316,7 @@ int main(int argc, char **argv)
 	timestatus_finish();
       }
 
-      /* imp3 is not negative, but fwrite needs an unsigned here */
-      fwrite(mp3buffer,1,(unsigned int)imp3,outf);
+      fwrite(mp3buffer,1,imp3,outf);
       lame_mp3_tags_fid(&gf,outf);       /* add ID3 version 1 or VBR tags to mp3 file */
       lame_close(&gf);
       fclose(outf);
